feat(sidebar): outline-first rail, Context hub with links/tags/AI subrail, toolbar Studio shortcut
Right-sidebar restructure (user direction 2026-07-16):

- Outline now leads the tab rail (order 10, ahead of everything)
- Standalone Links + Tags tabs MERGED into the Context tab as sub-tabs.
  Context renders an iconized subrail — Links · Tags · AI — with class
  parity to the left sidebar's file-tree sub-affordance row (same h-6
  bar, bridge strip, subActive/subInactive treatments). Context is now
  contentTypes "all" and no longer filtered by studio enablement: links
  and tags are core; only the AI sub-tab (agentic metadata, now
  ContextAiPanel) follows the studio extension toggle.
- Sub-tab availability: links → note/external (old backlinks scope),
  tags → all, AI → folder/note/file/html/code/external. Default lands
  on AI when available, else first available; explicit choice sticks.
- DEFAULT_RIGHT_SIDEBAR_TAB backlinks → outline; "backlinks"/"tags"
  stay in the union so persisted per-content tab choices still parse
  and fall back gracefully.
- ContentToolbar: Sparkles "Open Folder Studio" button, folders only
  (speed-reader button pattern) — expands the right panel if collapsed
  and jumps to the Studio tab.

Gate: typecheck clean · lint 0 errors (0 new) · build exit 0.
