+++ components/content/RightSidebar.tsx
@@ -21,7 +21,6 @@ import { useIsExtensionEnabled } from "@/lib/extensions/client-registry";
 import {
   STUDIO_EXTENSION_ID,
   STUDIO_TAB_KEY,
-  STUDIO_CONTEXT_TAB_KEY,
 } from "@/extensions/studio/manifest";
 import {
   resolveRightSidebarTab,
@@ -59,11 +58,9 @@ export function RightSidebar() {
     })
       .map((tool) => tool.tabKey)
       .filter(Boolean)
-      .filter(
-        (tabKey) =>
-          studioEnabled ||
-          (tabKey !== STUDIO_TAB_KEY && tabKey !== STUDIO_CONTEXT_TAB_KEY)
-      ) as RightSidebarTab[];
+      // Context stays regardless (its links/tags sub-tabs are core); only
+      // the Studio tab follows extension enablement.
+      .filter((tabKey) => studioEnabled || tabKey !== STUDIO_TAB_KEY) as RightSidebarTab[];
 
     // Properties tab is available when a block is selected (injected by RightSidebarHeader)
     if (selectedBlockId && !tabs.includes("properties")) {
